Fix issue with settings mixin from Fabric getting in the way (v1.8.1)
This is a problem for any components that do `from fabric.api import *`
as there's a `settings` context manager function that gets imported
making the `test` task think that it's inside a Django app.

This shouldn't be a problem for projects other than `armstrong.cli`
which is not a Django app.