change default MIDDLEWARE_CLASSES to use Django 1.7 default
This will require one change to the `env_settings.py` of some Armstrong components, probably only those that use the Admin though. Since the change only effects component developers?not end users?and is a minimal setting modification?not an interface change or feature removal?it's okay to do in a point release.